Clarifies standard api types for span
Prior to this change, there was a mutable accumulator called `Span` (similar to HTrace's one), and also a concept of a immutable finished span called `RawSpan`. Since `RawSpan` was the input of `RecordSpan`, and "raw" distracts at least me, this changes terminology to say `SpanRecord`. I believe `SpanRecord` more closely relates both the the apis that use it, and also it being a the canonical immutable struct.