Removes me from the immortals list
Last year, I got a lot of unsolicited support requests about OpenTracing for reasons including being the first person
on this list (and also accidentally ranking in the the opentracing-java repo due to misconfiguration). I'm removing
myself from this list as I need to cut down on the support work I'm doing.