edit opentracing.io license info so that GitHub recognizes it
GitHub uses a library called Licensee to identify a project's license
type. It shows this information in the status bar and via the API if it
can unambiguously identify the license.

This commit replaces the contents of the LICENSE file with the text
of the Apache 2.0 license (Copyright The OpenTracing Authors). It also
updates package.json to refer to the Apache license and slightly
modifies the header in the README. This all allows Licensee's license
type identification (as Apache 2.0) to succeed.

Signed-off-by: Andrea Kao <eirinikos@gmail.com>